clean conflict in TrappedGame.java AB#98
