+++ src/com/trapped/utilities/FileManager.java
@@ -6,7 +6,9 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Scanner;
+import java.util.stream.Stream;
 
 
 /*
@@ -26,6 +28,26 @@ public class FileManager {
         Files.copy(Path.of(art), out);
         out.flush();
     }
+    public static void readMessageSlowly(String fileName, int sec) throws InterruptedException {
+        String message = convertTxtToString(fileName);
+        char[] chars = message.toCharArray();
+        for (int i = 0; i < chars.length; i++) {
+            System.out.print(chars[i]);
+            Thread.sleep(sec);
+        }
+    }
+    public static String convertTxtToString(String fileName){
+        String file = "resources/" + fileName;
+        Path path = Paths.get(file);
+        StringBuilder sb = new StringBuilder();
 
+        try (Stream<String> stream = Files.lines(path)) {
+            stream.forEach(s -> sb.append(s).append("\n"));
+        } catch (IOException ex) {
+        }
+        String contents = sb.toString();
+        return contents;
+
+    }
 }
